Started working on assignment project. Imported a leaf model via Assimp, implemented hold/release based hover mechanics and a 'wind-like' sway in for the leaves in the vertex shader. Some code/comment cleanup.TODO: Setting lighting uniforms on shader switching needs to be refactored to a separate function as it currently causes huge code clutter.

+++ src/Camera.h
@@ -122,7 +122,7 @@ public:
     }
 
     void ProcessMouseScroll(float yOffset) {
-        Zoom -= (float) yOffset;
+        Zoom -= yOffset;
 
         if (Zoom < 1.0f)
             Zoom = 1.0f;
